+++ src/scenes/AuthSelect/AuthSelect.js
@@ -1,11 +1,15 @@
 import React, { useEffect } from 'react';
 import './AuthSelect.css';
 import Modal from '../../components/Modal/Modal';
+import analytics, { EVENT_NAMES } from 'utils/analytics';
 
 function AuthSelect({ onClose, login }) {
   useEffect(() => {
     console.log('Mounting auth UI...');
-    window.Virtru.AuthWidget('virtru-auth-widget-mount', { afterAuth: login });
+    analytics.track({ event: EVENT_NAMES.LOGIN_ATTEMPT });
+    window.Virtru.AuthWidget('virtru-auth-widget-mount', {
+      afterAuth: login,
+    });
   }, [login]);
 
   return (