+++ src/scenes/Document/scenes/DownloadModal/DownloadModal.js
@@ -3,14 +3,80 @@ import React, { useEffect, useState } from 'react';
 import Modal from 'components/Modal/Modal';
 import Button from 'components/Button/Button';
 import { downloadHtml, downloadTdf, downloadDecrypted } from 'utils/download';
+import { EVENT_NAMES, trackDownload } from 'utils/analytics';
 
 import './DownloadModal.css';
 import '../LoadingModal/LoadingModal.css';
 
 import PolicyUtils from 'utils/policy';
 
-export default ({ onClose, encrypted, virtruClient }) => {
+const DOWNLOAD_TYPES = {
+  FILE: 'file',
+  TDF: 'tdf',
+  TDF_HTML: 'tdf.html',
+};
+
+export default ({ onClose, encrypted, virtruClient, policy }) => {
   const [decrypting, setDecrypting] = useState(false);
+
+  /** Track Download Metrics for Amplitude */
+  const downloadAsHtml = () => {
+    trackDownload({
+      event: EVENT_NAMES.FILE_DOWNLOAD_ATTEMPT,
+      extension: DOWNLOAD_TYPES.TDF_HTML,
+      isSecure: true,
+      encrypted,
+      policy,
+    });
+    try {
+      downloadHtml(encrypted);
+      trackDownload({
+        event: EVENT_NAMES.FILE_DOWNLOAD_COMPLETE,
+        extension: DOWNLOAD_TYPES.TDF_HTML,
+        isSecure: true,
+        encrypted,
+        policy,
+      });
+    } catch (error) {
+      trackDownload({
+        event: EVENT_NAMES.FILE_DOWNLOAD_ERROR,
+        name: error.name,
+        message: error.message,
+        stack: error.stack,
+        encrypted,
+        policy,
+      });
+    }
+  };
+
+  const downloadAsTdf = () => {
+    trackDownload({
+      event: EVENT_NAMES.FILE_DOWNLOAD_ATTEMPT,
+      extension: DOWNLOAD_TYPES.TDF,
+      isSecure: true,
+      encrypted,
+      policy,
+    });
+    try {
+      downloadTdf(encrypted);
+      trackDownload({
+        event: EVENT_NAMES.FILE_DOWNLOAD_COMPLETE,
+        extension: DOWNLOAD_TYPES.TDF,
+        isSecure: true,
+        encrypted,
+        policy,
+      });
+    } catch (error) {
+      trackDownload({
+        event: EVENT_NAMES.FILE_DOWNLOAD_ERROR,
+        name: error.name,
+        message: error.message,
+        stack: error.stack,
+        encrypted,
+        policy,
+      });
+    }
+  };
 
   const showDecryptModal = () => {
     return (
@@ -29,12 +95,12 @@ export default ({ onClose, encrypted, virtruClient }) => {
           </h2>
           <span>Share with others:</span>
           <br />
-          <Button fullWidth onClick={() => downloadHtml(encrypted)}>
+          <Button fullWidth onClick={() => downloadAsHtml()}>
             Download HTML
           </Button>
           <span>Inspect the metadata:</span>
           <br />
-          <Button fullWidth onClick={() => downloadTdf(encrypted)}>
+          <Button fullWidth onClick={() => downloadAsTdf()}>
             Download TDF
           </Button>
         </Modal>
@@ -45,11 +111,21 @@ export default ({ onClose, encrypted, virtruClient }) => {
   const showDecryptAndDownloadModal = () => {
     const decryptAndDownload = async () => {
       setDecrypting(true);
+      trackDownload({ event: EVENT_NAMES.FILE_DOWNLOAD_ATTEMPT, encrypted, policy });
       try {
         await downloadDecrypted({ encrypted, virtruClient });
-      } catch (err) {
-        console.error(err);
+        trackDownload({ event: EVENT_NAMES.FILE_DOWNLOAD_COMPLETE, encrypted, policy });
+      } catch (error) {
+        console.error(error);
         alert('File could not be decrypted');
+        trackDownload({
+          event: EVENT_NAMES.FILE_DOWNLOAD_ERROR,
+          name: error.name,
+          message: error.message,
+          stack: error.stack,
+          encrypted,
+          policy,
+        });
       } finally {
         setDecrypting(false);
       }
@@ -61,12 +137,12 @@ export default ({ onClose, encrypted, virtruClient }) => {
           <h1>Download File</h1>
           <span>Share with others:</span>
           <br />
-          <Button fullWidth onClick={() => downloadHtml(encrypted)}>
+          <Button fullWidth onClick={() => downloadAsHtml()}>
             Download HTML
           </Button>
           <span>Inspect the metadata:</span>
           <br />
-          <Button fullWidth onClick={() => downloadTdf(encrypted)}>
+          <Button fullWidth onClick={() => downloadAsTdf()}>
             Download TDF
           </Button>
           <span>See the original file:</span>